"fix: Cambiar tipo de ID de Integer a Long en servicios de Producto"

+++ src/main/java/com/rodolfogodinez/kinalapp/service/ProductoService.java
@@ -13,7 +13,6 @@ public class ProductoService implements IProductoService {
 
     private final ProductoRepository productoRepository;
 
-    // Inyección por constructor
     public ProductoService(ProductoRepository productoRepository) {
         this.productoRepository = productoRepository;
     }
@@ -32,46 +31,34 @@ public class ProductoService implements IProductoService {
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<Producto> buscarPorCodigo(Integer codigo) {
+    public Optional<Producto> buscarPorCodigo(Long codigo) {
         return productoRepository.findById(codigo);
     }
 
     @Override
     public Producto guardar(Producto producto) {
-        // Validar datos del producto
         validarProducto(producto);
-
-        // Si no se especifica estado, se activa por defecto
         if (producto.getEstado() == null || producto.getEstado() == 0) {
             producto.setEstado(1);
         }
-
-        // Si no se especifica stock, se inicializa en 0
         if (producto.getStock() == null) {
             producto.setStock(0);
         }
-
         return productoRepository.save(producto);
     }
 
     @Override
-    public Producto actualizar(Integer codigo, Producto producto) {
-        // Verificar que el producto existe
+    public Producto actualizar(Long codigo, Producto producto) {
         if (!productoRepository.existsById(codigo)) {
             throw new RuntimeException("Producto no encontrado con código: " + codigo);
         }
-
-        // Asegurar que el código del objeto coincida con el de la URL
         producto.setCodigoProducto(codigo);
-
-        // Validar datos
         validarProducto(producto);
-
         return productoRepository.save(producto);
     }
 
     @Override
-    public void eliminar(Integer codigo) {
+    public void eliminar(Long codigo) {
         if (!productoRepository.existsById(codigo)) {
             throw new RuntimeException("Producto no encontrado con código: " + codigo);
         }
@@ -80,7 +67,7 @@ public class ProductoService implements IProductoService {
 
     @Override
     @Transactional(readOnly = true)
-    public boolean existePorCodigo(Integer codigo) {
+    public boolean existePorCodigo(Long codigo) {
         return productoRepository.existsById(codigo);
     }
 
@@ -91,30 +78,24 @@ public class ProductoService implements IProductoService {
     }
 
     @Override
-    public Producto actualizarStock(Integer codigo, Integer cantidadVendida) {
+    public Producto actualizarStock(Long codigo, Integer cantidadVendida) {
         Producto producto = productoRepository.findById(codigo)
                 .orElseThrow(() -> new RuntimeException("Producto no encontrado"));
-
         Integer nuevoStock = producto.getStock() - cantidadVendida;
-
         if (nuevoStock < 0) {
             throw new IllegalArgumentException("Stock insuficiente. Stock actual: " + producto.getStock());
         }
-
         producto.setStock(nuevoStock);
         return productoRepository.save(producto);
     }
 
-    // Método privado para validaciones
     private void validarProducto(Producto producto) {
         if (producto.getNombreProducto() == null || producto.getNombreProducto().trim().isEmpty()) {
             throw new IllegalArgumentException("El nombre del producto es obligatorio");
         }
-
         if (producto.getPrecio() == null || producto.getPrecio() <= 0) {
             throw new IllegalArgumentException("El precio debe ser mayor a 0");
         }
-
         if (producto.getStock() == null || producto.getStock() < 0) {
             throw new IllegalArgumentException("El stock no puede ser negativo");
         }
